開發購物車功能

+++ ModelApplication/app/src/main/java/com/example/modelapplication/OrderCartItem.java
@@ -6,7 +6,7 @@ public class OrderCartItem {
 
     public String name;
     public int price = 0;
-    public float offset = 0f;
+    public double offset = 0f;
     public int count = 0;
     public int MaxCount = 0;
 }
